refactor(catalog-cli): carve backups command family into catalog_cmds (nexus-whh61.4)
Continue the catalog_cmds/ decomposition past kgyoz (now closed). Carve the
backup-snapshot lifecycle commands (list-backups, undelete, vacuum-backups)
out of commands/catalog.py into commands/catalog_cmds/backups.py, mirroring
the established owners/backfill/links pattern: plain @click.command +
register(group); commands.catalog calls _backups_cmds.register(catalog).
'nx catalog list-backups' (etc.) resolve identically — names, options,
output unchanged; 47 total catalog commands preserved.

Behaviour-preserving: bodies moved verbatim (incl. lazy catalog_backup /
factory imports and undelete's make_catalog_admin admin path). list-backups
and vacuum-backups reach _get_catalog via the nexus.commands.catalog module
object inside the body (acyclic + monkeypatch-safe); undelete uses no shared
helper. No test imports these by symbol.

Three contract pins in test_catalog_cli.py::TestWhh61BackupsCarve:
registration, callbacks-in-carved-module, behavioural module-routed
_get_catalog via list-backups. Guard suites green (115: cli + backup-safety).
